requirements.txt
# Core Requirements
flask==2.3.2
flask-cors==5.0.1
gunicorn==23.0.0

# Deep Learning
tensorflow-cpu==2.12.0  # Using stable CPU version for cloud
deepface==0.0.93

# Image Processing
opencv-python-headless==4.10.0.84  # Headless version for server
pillow==9.5.0

# Numerical Computing
numpy==1.26.4

# Protocol Buffers
protobuf==3.20.3  # Prevents conflicts with TensorFlow